chore: tidy up, improve tests
